FAC_ACL_usability1
Users with multiple facilities selected may only see schedule for one
facility at a time because "All Facilities" is disabled.  This allows
them to select the "all facilities" choice in the calendar, yet only see
the combined schedule for their assigned facilities (not really all).
Also snuck in a little fix for patient.inc, because a semicolon snuck in
instead of a comma.  Bad ju-ju.