button_ros_round4
Fully tested, behavior much as desired.  Round 5 may be on a later
commit to reduce the amount of code by appx 75%.  This is a brute-force
form, and is therefore not up to standard.  Translations should happen
against arrays drawn from a configurable table of element names before
any other processing.