originally from epsdky appointment changes
I have written a description in case you want to pull in the above code

Extended Appointment Display Sets (Patient Summary)

Option added to Appointment Display Sets to allow (if necessary) the
appointment display limit to be disregarded so that all appointments can
be displayed for the last set.

for Appointment Display Sets appointments with same date are grouped
into sets for display.

The last set turns green (default) to indicate that all member
appointments are diplayed and at least one subsequent scheduled
appointment exists (not displayed).

The last set turns pink (default) when not all member appointments are
displayed.