dated reminders #144  user selection box ORDER, WHERE, and size change
Alphabetize the user dropdown list, ensure that the list contains only
real local users by checking some additional fields, make the select
show 8 instead of the default 5, and set the list of reminders to 15
instead of just 5.  One heavy user has theirs set to 50, so this is a
minimal compromise.