logo format and location
improved logo appearance and format.  Also cleaned up location of
resources and user-configurable logo locations as they appear in the
code.  Will make a note for new issue in next release that deals with
user-configuration at a larger level.  Sign-on page logo should be
rendered AGAIN from the SVG file that I don't have available, and the
layout of the sign-in frame needs to accomodate (scale) a range of image
sizes, or consume a specified size.