Changed autoload call to $GLOBALS['modules_dir']
This change breaks the branch, and will not be testable without merging
with master.  It was indicated that this would be a good idea, but not a
review request.  I assumed that meant "do a new commit after merge", but
I am treating it as a review change to see if this passes muster.