Documentation and installer helpers
Re-committing the changes to the documentation directory structure
posted to the old repo.
This commit organizes the documentation directory, and adds the
installer helper docs to assets.  The new files are utilized by the
XAMPP package install helper, and the compiled installer itself.  An
attempt was made to replace 1:1 files existing in the current master and
overwrite them into the new structure over the versions used in the
original branch last September.