Encounter Merging tool
Placed in /modules.  This is a WIP in the fact that it *may* need to be
added to the module list in globals and into the menu JSON to be useful.
It could also be integrated if really desired, or added to the
non-existent add-on tools/forms/reports repo.
Encounter merging tool is by slikins, written for SunCoast Connection