JS header include library function (#561)
This commit brought to you as an effort to push one tiny cleaning commit
every day.  This particular function had three includes, an additional
directory with a single file and a single function to create a path that
spanned up and down the tree so that two lines of code could be made a
little longer.