move js/datatables to assets
Datatables is used in several areas.  The biggest issue is in the tags
system where the assets are scattered everywhere, and there is an
additional singleton file called datatables.js in addition to this
package.  Those assets are too complex to handle/combine/move in a
single go, as there are additional copies of bootstrap and many other
packages in several directory layers.