Message notification added upper right corner
If there are unread messages, a notification will appear showing "5 unread messages" for example. If there are no unread messages, no notification will appear.

Messages notification have an absolute position with 50px right. This design isn't responsive and I think it's not a good practice, but the userData button was design this way, so I followed.

This notification ONLY refreshes when the page refreshes (opening, refreshing or closing tabs DO NOT refresh the page, therefore, it's completely predictable that the notification remains the same). There are probably workarounds to fix this.