Flowboard layout fix (#787)
* Test

Test

* Fixed the layout and checkbox repositioned

The "open patient in new window" checkbox was repositioned to its place. Now it is located on pat_settings, the icon next to Flow Board which opens a list of options.

Because of that, the layout was fixed.

* Removed this little @£@§£#$#$

* topatient() bug fixed

Since the input checkbox was changed to a different form, this function must be changed too, otherwise it would be looking for an input on the wrong place.

I believe it's fixed now.