Created a confirmation dialog, made changes and idented the code
The code was a mess and I idented it and now it must be fine.

Created a confirm message when clicking on submit to avoid accidental clicks. The message is now red.

The successful message is now an Alert() .

There is an issue on line 57. There must be ONLY one bracket at the beginning ( } ) and the code needs 2 of them to work. If i remove one of them, the tab is going to be Undefined and it just doesn't load. I really don't know why this is happening.