Changed library stuff in order to fix the layout issue
I changed some values of "top" properties directly on the library. I've set the popup "position" to "fixed" because I commented the code that was handling the positions when scrolling. That had to be made because when we scrolled, the page would become infinite. Besides, "position: fixed" exists for some purpose.