Tab buttons, bugfixes and more (#777)
* Changed tab icons, added tooltips and fixed bugs

A "}" was missing in style.css and I wrote it back there.
I created a new CSS class for tab buttons (close etc). Those buttons have now a margin-left for a better view.
Those buttons were changed to font-awesome 4.7.0. The task says that I should make this responsive between styles but I am unable to change the button class depending on the selected user theme (--> any suggestion is welcome).

I tried to add a tooltip made in HTML + CSS but it was hard to incorporate it in the code because it's hard to find a nice position for the tooltip to appear responsively to the screen and mouse coordinates. The best method I found was to use html tag property "title" which wasn't probably what the task proposed but I think it's pretty suitable anyway. If you guys really want me to do a tooltip in CSS, I'll need some guidance on how to do it (not the design itself because that's not really my problem) because I don't know how I can position the tooltip div and fit perfectly like html "title" does.

Added a dependency for the font-awesome on headers.inc.php and added its folder on assets but didn't remove the older version of font-awesome because it can be linked to something on the project and removing it would probably cause unexpected problems.

* Removed .gitignore

I dont really know what it is doing there but i think this commit can remove it

* Font-awesome version migration

Removed the new version asset and changed the dependency to the old one. Multiple font-awesome copies like jQuery's avoided.

* Changed calling of dependency

Now the dependency is called within "call_required_libraries".

* added newline

* fixed conflict