Idented javascript code
There is a problem here that is out of this PR's scope: livequery library is overriding jquery's $. That gives an error to the console. Tried to replace $ with jQuery() but didn't work
I removed livequery just for testing purposes and confirmed what i said, livequery is overriding jquery.

Did no changes