Access control based on Patient Facility
Feature addition to allow a clinic to restrict access to all patient
data based on the confluence of the `user` table schedule_facilities and
the `patient_data` 'facility' fields.

This feature allows multiple clinics to share a database, while
restricting whole popluations of patient data from view.  This has the
effect of both controlling access as well as simplifying the data for
users that have no need to see what is going on in other physical or
organizational clinical service locations.

This feature is of extreme importance for Behavioral Health,
Secure/institutional Care facilities, and Assisted Living Facility
organizations.  It is also very useful for temporary care provider
situations and many others.

This is the third attempt at implementing this feature.  Prior PR's
included data schema changes that were controversial, but those changes
have since been included in recent pull requests from other
contributors.  Rather than have the history of those discussions, this
branch was created fresh for simplification of review.