Clean up the code for the next set of changes (#1142)
* Clean up the code for the next set of changes

Clean up the code for the next set of changes

* Add endtime to provider side

Add endtime to provider side

* Removed commented out code

Removed commented out code

* Comment change per Art's request

Comment change per Art's request

* Art's typo

Art's typo

* Fix to code.

Need to re-visit the allday tags.

* Removal of the delete button on the patient side

Removal of the delete button on the patient side

* Added the cancel reason to the screen

Added the cancel reason to the screen for the patient side.

* Database changes for the cancellation reason

Database changes for the cancellation reason incremented the version
file. Added the delet appt status to the upgrade file.