fix: unstyled table (#1360) Layout Editor changes
Layout based demographics changes.  This corrects the arrangement of the fields (while not doing anything to make the form more explicable).  Form is totally much cleaner.  I tested in Prism, not Light, but Mua tested in both.  I assume that means testing the actual interface and not just how it presents!